(GP-2.3.1): Added Blob Verification and Reset Functionality (#4)
In this update, I added the verifyBlob method to programmatically confirm that blob files exist in the objects directory. The method computes the SHA-1 hash of given content and checks whether a corresponding blob file is present. I also implemented resetRepo in the tester to delete and reinitialize the repository, allowing blob creation and verification to be tested multiple times. These improvements make the system more robust and provide a reliable way to confirm blob functionality.